+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Push new releases to `capz-app-collection`
 
+### Changed
+
+- Decrease TTL for `A-Records` from 60 minutes down to 5 minutes
+
 ## [1.0.3] - 2023-03-09
 
 ### Changed