+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-03-28
+
 ### Added
 
 - Push new releases to `capz-app-collection`
@@ -69,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release
 
-[Unreleased]: https://github.com/giantswarm/dns-operator-azure/compare/v1.0.3...HEAD
+[Unreleased]: https://github.com/giantswarm/dns-operator-azure/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/giantswarm/dns-operator-azure/compare/v1.0.3...v1.1.0
 [1.0.3]: https://github.com/giantswarm/dns-operator-azure/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/giantswarm/dns-operator-azure/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/giantswarm/dns-operator-azure/compare/v1.0.0...v1.0.1